+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
-- Changed /setrung command to /rung, which now outputs the players current rung if no input is provided to set the rung.
+- Split /setrung command into /rung set and /rung get, which modify and output a player's echeladder rung respectively.
 
 ## [1.19.2-1.11.0.0] - 2023-07-22